sdk/pyproject.toml
[project]
name = "dginf"
version = "0.1.0"
description = "DGInf consumer SDK — private decentralized AI inference"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "httpx-sse>=0.4",
    "pydantic>=2.0",
    "typer>=0.15",
    "tomli>=2.0; python_version < '3.11'",
    "tomli-w>=1.0",
    "rich>=13.0",
    "PyNaCl>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
dginf = "dginf.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
